Create honey.py
This is a simple honeypot written in Python. Run this with nohup from a public facing directory on your server and it will log all traffic attempting to reach port 3389 (Remote Desktop Protocol). You can use the "Investigate Failed Login" scripts to cross reference with VirusTotal and get intel on who is sniffing around for vulnerabilities on your server.